requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
google-genai==0.3.0
python-dotenv==1.0.0
